+++ store/user.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const loginStateAtom = atom<boolean>({
+  key: 'loginStateAtom',
+  default: false,
+});
